review feedback applied
