+++ src/community/features-templating/features-templating-core/src/main/java/org/geoserver/featurestemplating/builders/impl/DynamicIncludeFlatBuilder.java
@@ -3,13 +3,8 @@ package org.geoserver.featurestemplating.builders.impl;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Map;
-import org.geoserver.featurestemplating.builders.TemplateBuilder;
-import org.geoserver.featurestemplating.builders.TemplateBuilderMaker;
 import org.geoserver.featurestemplating.builders.visitors.TemplateVisitor;
-import org.geoserver.featurestemplating.readers.JSONTemplateReader;
-import org.geoserver.featurestemplating.readers.TemplateReaderConfiguration;
 import org.geoserver.featurestemplating.writers.TemplateOutputWriter;
 import org.xml.sax.helpers.NamespaceSupport;
 
@@ -42,21 +37,6 @@ public class DynamicIncludeFlatBuilder extends DynamicValueBuilder {
         if (evaluate instanceof JsonNode && hasDynamic((JsonNode) evaluate))
             writeFromNestedTree(context, writer, (JsonNode) evaluate);
         else writeValue(writer, evaluate, context);
-    }
-
-    private void writeFromNestedTree(
-            TemplateBuilderContext context, TemplateOutputWriter writer, JsonNode node)
-            throws IOException {
-        TemplateReaderConfiguration configuration =
-                new TemplateReaderConfiguration(getNamespaces());
-        JSONTemplateReader jsonTemplateReader =
-                new JSONTemplateReader(node, configuration, new ArrayList<>());
-        TemplateBuilderMaker maker = configuration.getBuilderMaker();
-        maker.namespaces(configuration.getNamespaces());
-        writer.startObject(getKey(context), getEncodingHints());
-        jsonTemplateReader.getBuilderFromJson(getKey(context), node, this, maker);
-        for (TemplateBuilder child : getChildren()) child.evaluate(writer, context);
-        writer.endObject(getKey(context), encodingHints);
     }
 
     private boolean hasDynamic(JsonNode node) {
