+++ src/community/features-templating/features-templating-core/src/main/java/org/geoserver/featurestemplating/builders/impl/DynamicMergeBuilder.java
@@ -3,14 +3,9 @@ package org.geoserver.featurestemplating.builders.impl;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.logging.Logger;
-import org.geoserver.featurestemplating.builders.TemplateBuilder;
-import org.geoserver.featurestemplating.builders.TemplateBuilderMaker;
 import org.geoserver.featurestemplating.builders.visitors.TemplateVisitor;
 import org.geoserver.featurestemplating.readers.JSONMerger;
-import org.geoserver.featurestemplating.readers.JSONTemplateReader;
-import org.geoserver.featurestemplating.readers.TemplateReaderConfiguration;
 import org.geoserver.featurestemplating.writers.TemplateOutputWriter;
 import org.geotools.util.logging.Logging;
 import org.xml.sax.helpers.NamespaceSupport;
@@ -87,21 +82,6 @@ public class DynamicMergeBuilder extends DynamicValueBuilder {
             else // contains static content
             writeValue(writer, mergedNodes, context);
         }
-    }
-
-    private void writeFromNestedTree(
-            TemplateBuilderContext context, TemplateOutputWriter writer, JsonNode node)
-            throws IOException {
-        TemplateReaderConfiguration configuration =
-                new TemplateReaderConfiguration(getNamespaces());
-        JSONTemplateReader jsonTemplateReader =
-                new JSONTemplateReader(node, configuration, new ArrayList<>());
-        TemplateBuilderMaker maker = configuration.getBuilderMaker();
-        maker.namespaces(configuration.getNamespaces());
-        writer.startObject(getKey(context), getEncodingHints());
-        jsonTemplateReader.getBuilderFromJson(getKey(context), node, this, maker);
-        for (TemplateBuilder child : getChildren()) child.evaluate(writer, context);
-        writer.endObject(getKey(context), encodingHints);
     }
 
     private boolean hasDynamic(JsonNode node) {
